ibc: remove component tree and move createclient to action handler
