+++ app/src/action_handler/actions/ibc_action/msg/create_client.rs
@@ -2,16 +2,19 @@ use std::sync::Arc;
 
 use anyhow::Result;
 use async_trait::async_trait;
+use ibc::core::ics02_client::client_state::ClientState;
 use ibc::core::ics02_client::msgs::create_client::MsgCreateClient;
+use ibc::core::ics24_host::identifier::ClientId;
 use penumbra_storage::{StateRead, StateWrite};
 use penumbra_transaction::Transaction;
 
 use crate::action_handler::ActionHandler;
+use crate::ibc::client::ics02_validation;
 use crate::ibc::component::client::{
-    stateful::create_client::CreateClientCheck,
     stateless::create_client::{client_state_is_tendermint, consensus_state_is_tendermint},
-    Ics2ClientExt as _,
+    StateReadExt as _, StateWriteExt as _,
 };
+use crate::ibc::{event, ClientCounter};
 
 #[async_trait]
 impl ActionHandler for MsgCreateClient {
@@ -27,11 +30,45 @@ impl ActionHandler for MsgCreateClient {
         Ok(())
     }
 
+    // execute IBC CreateClient.
+    //
+    //  we compute the client's ID (a concatenation of a monotonically increasing integer, the
+    //  number of clients on Penumbra, and the client type) and commit the following to our state:
+    // - client type
+    // - consensus state
+    // - processed time and height
     async fn execute<S: StateWrite>(&self, mut state: S) -> Result<()> {
         tracing::debug!(msg = ?self);
-        state.validate(self).await?;
-        state.execute_create_client(self).await?;
+        let client_state =
+            ics02_validation::get_tendermint_client_state(self.client_state.clone())?;
 
+        // get the current client counter
+        let id_counter = state.client_counter().await?;
+        let client_id = ClientId::new(client_state.client_type(), id_counter.0)?;
+
+        tracing::info!("creating client {:?}", client_id);
+
+        let consensus_state =
+            ics02_validation::get_tendermint_consensus_state(self.consensus_state.clone())?;
+
+        // store the client data
+        state.put_client(&client_id, client_state.clone());
+
+        // store the genesis consensus state
+        state
+            .put_verified_consensus_state(
+                client_state.latest_height(),
+                client_id.clone(),
+                consensus_state,
+            )
+            .await
+            .unwrap();
+
+        // increment client counter
+        let counter = state.client_counter().await.unwrap_or(ClientCounter(0));
+        state.put_client_counter(ClientCounter(counter.0 + 1));
+
+        state.record(event::create_client(client_id, client_state));
         Ok(())
     }
 }
